requirements.txt
yfinance
yahoo_fin.stock_info

pandas
datetime

altair
streamlit

